[Feat][Lex] Refactored MainLayout to EmployeeLayout
