+++ client/src/App.jsx
@@ -2,31 +2,26 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
-import Home from "@components/pages/Home";
+import { ProtectedRoutes } from "@components/module";
 import Login from "@components/pages/Login";
-import Order from "@components/pages/Order";
-import Product from "@components/pages/Product";
-import Supplier from "@components/pages/Supplier";
-import { AuthProvider, useAuth } from "@hooks/Auth";
-
-function LoggedIn() {
-  const [auth] = useAuth();
-  return <h1>{`LOGGED IN! ${auth.first_name} ${auth.last_name}`}</h1>;
-}
+import EmployeeLayout from "@components/template/EmployeeLayout";
+import { AuthProvider } from "@hooks/Auth";
 
 function App() {
   return (
     <AuthProvider>
       <Router>
         <Routes>
-          {/*<Route path="/" element={<ProtectedRoutes />}>*/}
-          {/*  <Route index element={<LoggedIn />} />*/}
-          {/*</Route>*/}
+          <Route path="/" element={<ProtectedRoutes />}>
+            <Route element={<EmployeeLayout />}>
+              <Route index element={<h1>Home</h1>} />
+              <Route path="/order" element={<h1>Order</h1>} />
+              <Route path="/product" element={<h1>Products</h1>} />
+              <Route path="/supplier" element={<h1>Supplier</h1>} />
+            </Route>
+          </Route>
           <Route path="/login" element={<Login />} />
-          <Route path="/" element={<Home />} />
-          <Route path="/order" element={<Order />} />
-          <Route path="/product" element={<Product />} />
-          <Route path="/supplier" element={<Supplier />} />
+          <Route path="*" element={<h1>404</h1>} />
         </Routes>
       </Router>
