--- client/src/components/layouts/MainLayout.module.css
@@ -1,5 +1,0 @@
-.layout {
-  display: flex;
-  background-color: #1d2834;
-  min-height: 100vh;
-}
