+++ client/src/components/module/Nav/index.js
@@ -1,0 +1,7 @@
+import Bar from "./NavBar";
+import Link from "./NavLink";
+
+export default {
+  Bar,
+  Link,
+};
