--- client/src/components/module/NavBar/index.js
@@ -1,1 +1,0 @@
-export { default as default } from "./NavBar";
