+++ client/src/components/module/index.js
@@ -1,2 +1,3 @@
 export { default as LoginForm } from "./LoginForm";
+export { default as Nav } from "./Nav";
 export { default as ProtectedRoutes } from "./ProtectedRoutes";
