--- client/src/components/pages/Home.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MainLayout from "../layouts/MainLayout";
-
-function Home() {
-    return (
-        <MainLayout>
-            Hello! Welcome to Pan
-        </MainLayout>
-    )
-}
-
-export default Home
