--- client/src/components/pages/Order.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MainLayout from "../layouts/MainLayout";
-
-function Order() {
-    return (
-        <MainLayout>
-            Order Page
-        </MainLayout>
-    )
-}
-
-export default Order
