--- client/src/components/pages/Product.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MainLayout from "../layouts/MainLayout";
-
-function Product() {
-    return (
-        <MainLayout>
-            Product Page
-        </MainLayout>
-    )
-}
-
-export default Product
