--- client/src/components/pages/Supplier.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import MainLayout from "../layouts/MainLayout";
-
-function Supplier() {
-    return (
-        <MainLayout>
-            Supplier Page
-        </MainLayout>
-    )
-}
-
-export default Supplier
