+++ client/src/components/template/EmployeeLayout/EmployeeLayout.module.css
@@ -1,0 +1,5 @@
+.container {
+  display: flex;
+  background-color: var(--primary-color);
+  min-height: 100vh;
+}
